🐵🐘 Checkpoint ./src/client.jsx:8/2991 ./src/utils/sentry.js:8/1611
