+++ src/client.jsx
@@ -1,36 +1,9 @@
-/* globals EDITOR_URL, ENVIRONMENT, PROJECT_DOMAIN */
+/* globals EDITOR_URL */
 import './polyfills.js';
-import * as Sentry from '@sentry/browser';
 import React from 'react';
 import {render} from 'react-dom';
 import App from './app.jsx';
-
-
-// First things first -- let's bring our error collection online:
-try {
-  Sentry.init({
-    dsn: 'https://4f1a68242b6944738df12eecc34d377c@sentry.io/1246508',
-    environment: ENVIRONMENT,
-    beforeSend(event) {
-      const json = JSON.stringify(event);
-      const scrubbedJSON = json.replace(/"persistentToken":"[^"]+"/g, `"persistentToken":"****"`);
-      const scrubbedEvent = JSON.parse(scrubbedJSON);
-
-      return scrubbedEvent;
-    },
-  });
-  
-  Sentry.configureScope((scope) => {
-    scope.setTag("bootstrap", "true");
-    scope.setTag("PROJECT_DOMAIN", PROJECT_DOMAIN);
-  });
-  
-  // Expose for use on the developer console:
-  window.Sentry = Sentry;
-} catch (error) {
-  console.warn("Error bringing Sentry online", error);
-}
-
+import {configureScope} from './utils/sentry';
 
 // Here's a bunch of browser support tests
 // If any of them don't work we can't run in this browser
@@ -53,6 +26,10 @@ window.bootstrap = () => {
     window.location.replace(EDITOR_URL + window.location.hash);
     return;
   }
+  
+  configureScope((scope) => {
+    scope.setTag("bootstrap", "true");
+  });
     
   const dom = document.createElement('div');
   document.body.appendChild(dom);
